Removed work done in issue-#2 Git workflow was disrupted.
